Refactoring make more structured
- Use base class logger and ssh instance
- Move remove unnecessary folders
- Move subcommands execution to relative files
