+++ lib/chagall/cli.rb
@@ -2,13 +2,21 @@
 
 require "clamp"
 require_relative "settings"
-require_relative "deploy/main"
-require_relative "compose/main"
+require_relative "deploy"
+require_relative "compose"
+require_relative "rollback"
 
 Clamp.allow_options_after_parameters = true
 
 module Chagall
   class Cli < Clamp::Command
+
+    def run(arguments)
+      parse(arguments)
+      Chagall::Settings.configure(collect_options_hash)
+      execute
+    end
+ 
     def self.options_from_config_file
       @options_from_config_file ||= begin
         config_path = File.join(Dir.pwd, "chagall.yml") || File.join(Dir.pwd, "chagall.yaml")
@@ -29,7 +37,7 @@ module Chagall
                environment_variable: opt[:environment_variable]
       elsif opt[:proc].is_a?(Proc)
         option opt[:flags],
-               opt[:environment_variable].gsub("CHAGALL_"),
+               opt[:environment_variable].gsub("CHAGALL_", ""),
                opt[:description],
                default: options_from_config_file[opt[:key]] || opt[:default],
                environment_variable: opt[:environment_variable] do |value|
@@ -37,7 +45,7 @@ module Chagall
         end
       else
         option opt[:flags],
-               opt[:environment_variable].gsub("CHAGALL_"),
+               opt[:environment_variable].gsub("CHAGALL_", ""),
                opt[:description],
                default: options_from_config_file[opt[:key]] || opt[:default],
                environment_variable: opt[:environment_variable]
@@ -49,61 +57,10 @@ module Chagall
       exit(0)
     end
 
-    subcommand "deploy", "Deploy the application to the server" do
-      def execute
-        binding.irb
-        Chagall::Settings.configure(collect_options_hash)
-
-        Chagall::Deploy::Main.new
-      end
-    end
-
-    subcommand "setup", "Setup the server for deployment" do
-      def execute
-        Chagall::Settings.configure(collect_options_hash)
-        Chagall::Setup::Main.new
-      end
-    end
-
-    subcommand "compose", "Run Docker Compose commands with arguments passed through" do
-      # Override parse method to handle all arguments after the subcommand
-      def parse(arguments)
-        if arguments.empty?
-          puts "ERROR: Missing required arguments"
-          puts "Usage: chagall compose COMMAND [OPTIONS]"
-          exit(1)
-        end
-
-        # Extract the first argument as command
-        @command = arguments.shift
-
-        # Store the rest as raw args
-        @raw_args = arguments
-
-        # Validate required arguments
-        if @command.nil? || @command.empty?
-          puts "ERROR: Command is required"
-          puts "Usage: chagall compose COMMAND [OPTIONS]"
-          exit(1)
-        end
-      end
-
-      def execute
-        Chagall::Settings.configure(collect_options_hash)
-        Chagall::Compose::Main.new(@command, @raw_args)
-      end
-    end
-
-    subcommand "rollback", "Rollback to previous deployment" do
-      option [ "--steps" ], "STEPS", "Number of steps to rollback", default: "1" do |s|
-        Integer(s)
-      end
-
-      def execute
-        Chagall::Settings.configure(collect_options_hash)
-        puts "Rollback functionality not implemented yet"
-      end
-    end
+    subcommand "deploy", "Deploy the application to the server", Chagall::Deploy
+    subcommand "setup", "Setup the server for deployment", Chagall::Setup
+    subcommand "compose", "Run Docker Compose commands with arguments passed through", Chagall::Compose
+    subcommand "rollback", "Rollback to previous deployment", Chagall::Rollback
 
     private
